code/pyproject.toml
[build-system]
requires = [
    "setuptools>=42",
    "wheel"
]
build-backend = "setuptools.build_meta"

[project]
name = "sqic"
version = "0.1.0"
description = "Lossy Image Compression with Stochastic Quantization"
authors = [
    {name = "Vladimir Norkin", email = "v.norkin@kpi.ua"},
    {name = "Anton Kozyriev", email = "a.kozyriev@kpi.ua"},
]
readme = "README.md"
license = {file = "LICENSE.code.md"}
dependencies = [
    "numpy>=1.26.4,<2",
    "pillow>=10.4.0,<2",
]
requires-python = ">=3.8"
keywords = [
    "non-convex optimization",
    "stochastic quantization",
    "color quantization",
    "lossy compression",
]
classifiers = [
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Graphics :: Editors :: Raster-Based",
]

[project.urls]
Homepage = "https://github.com/kaydotdev/sqic"
Issues = "https://github.com/kaydotdev/sqic/issues"
Repository = "https://github.com/kaydotdev/sqic.git"
